requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
pandas==2.1.4
gspread==5.12.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
openai==0.28.1
python-dotenv==1.0.0
lxml==4.9.3
openpyxl==3.1.2
schedule==1.2.0
